# api-contracts/poc-api-contracts.yaml
openapi: 3.0.3
info:
  title: CBS BI Platform POC API
  version: 1.0.0
  description: |
    Documentation subset for the public POC endpoints used by the React client.
    Runtime ASP.NET Core Swagger remains the authoritative generated contract.
servers:
  - url: https://localhost:7116
    description: Local development

paths:
  /api/Analytics/ask:
    post:
      summary: Ask an analytics question in natural language
      operationId: askAnalyticsQuestion
      tags: [Analytics]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalyticsQuestionRequest'
      responses:
        '200':
          description: Generated SQL and analytics result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsResponse'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          description: Analytics query could not be generated safely from available semantic context
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '504':
          description: Analytics request exceeded the configured execution timeout
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'

  /api/analytics/visual-query:
    post:
      summary: Execute a structured visual analytics query
      operationId: executeVisualQuery
      tags: [Visual Query]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VisualQueryRequest'
      responses:
        '200':
          description: Deterministically generated SQL and analytics result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsResponse'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          description: Unsupported or invalid visual query definition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'

  /api/analytics/saved-queries:
    get:
      summary: List saved queries for the current user
      operationId: getSavedQueries
      tags: [Saved Queries]
      responses:
        '200':
          description: Saved queries owned by the current user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SavedAnalyticsQuery'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      summary: Save a business analytics question
      operationId: saveAnalyticsQuery
      tags: [Saved Queries]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveAnalyticsQueryRequest'
      responses:
        '200':
          description: Saved query
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavedAnalyticsQuery'
        '201':
          description: Saved query created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavedAnalyticsQuery'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/Forbidden'

  /api/analytics/saved-queries/{id}:
    delete:
      summary: Delete a saved query owned by the current user
      operationId: deleteSavedQuery
      tags: [Saved Queries]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Deleted
        '204':
          description: Deleted
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/Forbidden'

  /api/analytics/dashboards:
    get:
      summary: List dashboards for the current user
      operationId: getDashboards
      tags: [Dashboards]
      responses:
        '200':
          description: Dashboards owned by the current user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AnalyticsDashboard'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      summary: Create a dashboard from saved analytics queries
      operationId: createDashboard
      tags: [Dashboards]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAnalyticsDashboardRequest'
      responses:
        '200':
          description: Created dashboard
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsDashboard'
        '201':
          description: Created dashboard
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsDashboard'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          description: Invalid dashboard definition or inaccessible saved query reference
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'

  /api/analytics/dashboards/{id}:
    get:
      summary: Get a dashboard by id
      operationId: getDashboardById
      tags: [Dashboards]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Dashboard definition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsDashboard'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Dashboard not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'

    delete:
      summary: Delete a dashboard owned by the current user
      operationId: deleteDashboard
      tags: [Dashboards]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: Deleted
        '204':
          description: Deleted
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/Forbidden'

components:
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Resource identifier returned by the backend

  responses:
    AuthenticationRequired:
      description: Authentication is required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'

    Forbidden:
      description: Authenticated user is not authorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'

  schemas:
    AnalyticsQuestionRequest:
      type: object
      required: [question]
      properties:
        question:
          type: string
          example: איזו עיר בעלת שיעור האבטלה הנמוך ביותר?

    AnalyticsResponse:
      type: object
      required: [sql, result]
      properties:
        sql:
          type: string
        result:
          $ref: '#/components/schemas/AnalyticsResult'

    AnalyticsResult:
      type: object
      required: [columns, rows]
      properties:
        columns:
          type: array
          items:
            type: string
        rows:
          type: array
          items:
            type: object
            additionalProperties: true

    VisualQueryRequest:
      type: object
      required: [domain, metric, sortDirection, limit]
      properties:
        domain:
          type: string
          enum:
            - Population
            - Employment
            - Housing
            - Education
        metric:
          type: string
          description: Metric must be valid for the selected domain
          example: Population
        year:
          type: integer
          nullable: true
          example: 2025
          description: Null means latest available year
        sortDirection:
          type: string
          enum:
            - Ascending
            - Descending
        limit:
          type: integer
          minimum: 1
          maximum: 100
          example: 5

    SaveAnalyticsQueryRequest:
      type: object
      required: [name, question]
      properties:
        name:
          type: string
          example: Lowest unemployment
        question:
          type: string
          example: איזו עיר בעלת שיעור האבטלה הנמוך ביותר?

    SavedAnalyticsQuery:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        question:
          type: string
        createdByUserId:
          type: string
          readOnly: true
        createdAtUtc:
          type: string
          format: date-time
        updatedAtUtc:
          type: string
          format: date-time

    VisualizationType:
      type: string
      enum:
        - Table
        - Number
        - BarChart

    CreateAnalyticsDashboardWidgetRequest:
      type: object
      required: [title, savedQueryId, visualizationType, displayOrder]
      properties:
        title:
          type: string
        savedQueryId:
          type: string
        visualizationType:
          $ref: '#/components/schemas/VisualizationType'
        displayOrder:
          type: integer
          minimum: 1

    AnalyticsDashboardWidget:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        savedQueryId:
          type: string
        visualizationType:
          $ref: '#/components/schemas/VisualizationType'
        displayOrder:
          type: integer

    CreateAnalyticsDashboardRequest:
      type: object
      required: [name, isDynamic, widgets]
      properties:
        name:
          type: string
          example: City Overview
        isDynamic:
          type: boolean
          example: false
        widgets:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/CreateAnalyticsDashboardWidgetRequest'

    AnalyticsDashboard:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        createdByUserId:
          type: string
          readOnly: true
        isDynamic:
          type: boolean
        widgets:
          type: array
          items:
            $ref: '#/components/schemas/AnalyticsDashboardWidget'
        createdAtUtc:
          type: string
          format: date-time
        updatedAtUtc:
          type: string
          format: date-time

    ProblemDetails:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string